+++ src/pages/Click.jsx
