+++ src/pages/Home.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export default function Home() {
+  return (
+    <section className="h-screen grid">
+      <div className="m-auto text-center text-3xl">
+        Welcome to My react beginners Projects!!
+      </div>
+    </section>
+  );
+}
